Homework3/requirements.txt
trl
transformers[torch]
datasets
wandb
python-dotenv
scikit-learn
deepspeed>=0.9.3
accelerate


